Host AbortableComparator in its own file.
Also include AbandonSort in AbortableComparator (which is the only
place that uses it).
