+++ jena-arq/src/main/java/org/apache/jena/atlas/data/SortedDataBag.java
@@ -82,76 +82,6 @@ public class SortedDataBag<E> extends AbstractDataBag<E>
         this.comparator = new AbortableComparator<E>(comparator);
     }
     
-    private static final class AbortableComparator<E> implements Comparator<E> 
-    	{
-    	/**
-    	    The test for whether the sort has been cancelled is
-    	    performed every <code>cancelTestFrequency</code> comparisons.
-    	    This reduces the (presumed) overhead of access to a
-    	    volatile boolean.    	    
-    	*/
-    	static final int cancelTestFrequency = 10000;
-    	
-    	/**
-    	    Count of the number of times this comparator has been called.
-    	*/
-		int count = 0;
-		
-		protected volatile boolean cancelled;
-		
-		final Comparator<? super E> baseComparator;
-		
-		public AbortableComparator(Comparator<? super E> comparator) 
-		{
-			this.baseComparator = comparator;
-		}
-
-		@Override public int compare(E o1, E o2) 
-		{	
-			count += 1;
-			if (count % cancelTestFrequency == 0) 
-			{
-				if (cancelled) throw new AbandonSort();
-			}
-			return baseComparator.compare(o1, o2);
-		}
-		
-		/**
-		    Sort the array <code>e</code> using this comparator
-		 	with the additional ability to abort the sort.
-		*/
-		public boolean abortableSort(E[] e) 
-		{
-			try 
-			{ 
-				Arrays.sort(e, this); 
-			}
-			catch (AbandonSort s) 
-			{ 
-				return true;
-			}
-			return false;
-		}
-		
-		/**
-		    Arrange that the next on-frequency cancellation test
-		    in compare will succeed, aborting the sort. 
-		*/
-		public void cancel() 
-		{
-			cancelled = true;
-		}
-	}
-
-    /**
-        <code>AbandonSort</code> is the exception thrown from
-        <code>AbortableComparator</code> to abandon a sort.
-    */
-	public static class AbandonSort extends RuntimeException 
-    {
-		private static final long serialVersionUID = 1L;
-    }
-    
 	/**
 	    cancel arranges that further comparisons using the supplied
 	    comparator will abandon the sort in progress.
